matcher/multimatch.cc: Fix warning from newer versions of GCC.
git-svn-id: file:///home/xapian-svn/xapian/trunk@6068 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/matcher/multimatch.cc
@@ -231,7 +231,7 @@ MultiMatch::MultiMatch(const Xapian::Database &db_, const Xapian::Query::Interna
     } else {
 	Xapian::doccount number_of_leaves = db.internal.size();
 	// Can't just use resize - that creates N copies of the same RSet!
-	for (int i = 0; i < number_of_leaves; ++i) {
+	for (size_t i = 0; i < number_of_leaves; ++i) {
 	    subrsets.push_back(Xapian::RSet());
 	}
